refactor: 에러코드 타입수정 및 유저와 상품 연관관계 매핑
사용자와 상품간의 연관관계를 수정하고 상품에서 발생하는 에러코드 추가 및 반영

Resolves:

+++ backend/src/main/java/com/campforest/backend/product/controller/ProductController.java
@@ -8,6 +8,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.Authentication;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,6 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.campforest.backend.common.ApiResponse;
+import com.campforest.backend.common.ErrorCode;
 import com.campforest.backend.config.s3.S3Service;
 import com.campforest.backend.product.dto.ProductDetailDto;
 import com.campforest.backend.product.dto.ProductRegistDto;
@@ -29,6 +31,8 @@ import com.campforest.backend.product.model.Category;
 import com.campforest.backend.product.model.Product;
 import com.campforest.backend.product.model.ProductType;
 import com.campforest.backend.product.service.ProductService;
+import com.campforest.backend.user.model.Users;
+import com.campforest.backend.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -42,23 +46,31 @@ public class ProductController {
 
 	private final ProductService productService;
 	private final S3Service s3Service;
+	private final UserService userService;
 
 	//게시물 작성
 	@PostMapping(consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
-	public ApiResponse<?> createProduct(
+	public ApiResponse<?> createProduct(Authentication authentication,
 		@RequestPart(value = "files", required = false) MultipartFile[] files,
 		@RequestPart(value = "productRegistDto") ProductRegistDto productRegistDto
-	) throws IOException {
+	) throws Exception {
+
+		Users users = userService.findByEmail(authentication.getName())
+			.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
 
 		List<String> imageUrls = new ArrayList<>();
-		for (MultipartFile file : files) {
-			String extension = file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf("."));
-			String fileUrl = s3Service.upload(file.getOriginalFilename(), file, extension);
-			imageUrls.add(fileUrl);
-			log.info("Uploaded file URL: " + fileUrl);
+		try {
+			for (MultipartFile file : files) {
+				String extension = file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf("."));
+				String fileUrl = s3Service.upload(file.getOriginalFilename(), file, extension);
+				imageUrls.add(fileUrl);
+			}
+		} catch (IOException e) {
+			return ApiResponse.createError(ErrorCode.PRODUCT_CREATION_FAILED);
 		}
 
 		productRegistDto.setProductImageUrl(imageUrls);
+		productRegistDto.setUserId(users.getUserId());
 
 		productService.createProduct(productRegistDto);
 
@@ -68,7 +80,12 @@ public class ProductController {
 	//게시물 정보 가져오기
 	@GetMapping
 	public ApiResponse<?> getProduct(@RequestParam Long productId) {
-		ProductDetailDto findProduct = productService.getProduct(productId);
+		ProductDetailDto findProduct;
+		try {
+			findProduct = productService.getProduct(productId);
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.PRODUCT_NOT_FOUND);
+		}
 
 		return ApiResponse.createSuccess(findProduct, "게시물 조회 성공하였습니다.");
 	}
@@ -76,37 +93,76 @@ public class ProductController {
 	//게시물 수정
 	@PutMapping(consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
 	public ApiResponse<?> updateProduct(
+		Authentication authentication,
 		@RequestPart(value = "files", required = false) MultipartFile[] files,
 		@RequestPart(value = "productUpdateDto") ProductUpdateDto productUpdateDto
-	) throws IOException {
+	) throws Exception {
+
+		Users user = userService.findByEmail(authentication.getName())
+			.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
+		Product product = productService.getProductById(productUpdateDto.getProductId())
+			.orElseThrow(() -> new Exception("제품 정보를 찾을 수 없습니다."));
+
+		if (!product.getUserId().equals(user.getUserId())) {
+			return ApiResponse.createError(ErrorCode.INVALID_AUTHORIZED);
+		}
 
 		List<String> imageUrls = new ArrayList<>();
 		if (files != null) {
-			for (MultipartFile file : files) {
-				String extension = file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf("."));
-				String fileUrl = s3Service.upload(file.getOriginalFilename(), file, extension);
-				imageUrls.add(fileUrl);
+			try {
+				for (MultipartFile file : files) {
+					String extension = file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf("."));
+					String fileUrl = s3Service.upload(file.getOriginalFilename(), file, extension);
+					imageUrls.add(fileUrl);
+				}
+			} catch (IOException e) {
+				return ApiResponse.createError(ErrorCode.PRODUCT_UPDATE_FAILED);
 			}
 		}
-
+		productUpdateDto.setUserId(user.getUserId());
 		productUpdateDto.setProductImageUrl(imageUrls);
-		productService.updateProduct(productUpdateDto.getProductId(), productUpdateDto);
+		try {
+			productService.updateProduct(productUpdateDto.getProductId(), productUpdateDto);
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.PRODUCT_UPDATE_FAILED);
+		}
 
 		return ApiResponse.createSuccessWithNoContent("게시물 수정에 성공하였습니다");
 	}
 
 	//게시물 삭제
 	@DeleteMapping()
-	public ApiResponse<?> deleteProduct(@RequestParam Long productId) {
-		productService.deleteProduct(productId);
+	public ApiResponse<?> deleteProduct(Authentication authentication, @RequestParam Long productId) throws Exception {
+
+		Users user = userService.findByEmail(authentication.getName())
+			.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
+		Product product = productService.getProductById(productId)
+			.orElseThrow(() -> new Exception("제품 정보를 찾을 수 없습니다."));
+
+		if (!product.getUserId().equals(user.getUserId())) {
+			return ApiResponse.createError(ErrorCode.INVALID_AUTHORIZED);
+		}
+
+		try {
+			productService.deleteProduct(productId);
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.PRODUCT_DELETION_FAILED);
+		}
 
 		return ApiResponse.createSuccessWithNoContent("게시물 삭제에 성공하였습니다");
 	}
 
+
 	//이미지 삭제
 	@DeleteMapping("/image")
 	public ApiResponse<?> deleteProduct(@RequestParam Long productImageId, @RequestPart Long productId) {
-		productService.deleteProductImage(productId, productImageId);
+		try {
+			productService.deleteProductImage(productId, productImageId);
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.PRODUCT_DELETION_FAILED);
+		}
 
 		return ApiResponse.createSuccessWithNoContent("삭제되었습니다");
 	}
@@ -125,17 +181,21 @@ public class ProductController {
 
 		Category categoryEnum = null;
 		if (category != null) {
-			System.out.println(category);
 			try {
 				categoryEnum = Category.valueOf(category);
 			} catch (IllegalArgumentException e) {
-				return ApiResponse.createError("잘못된 카테고리 값입니다.");
+				return ApiResponse.createError(ErrorCode.INVALID_PRODUCT_CATEGORY);
 			}
 		}
 
 		Pageable pageable = PageRequest.of(page, size);
-		Page<ProductSearchDto> result = productService.findProductsByDynamicConditions(categoryEnum,
-			productType, minPrice, maxPrice, locations, titleKeyword, pageable);
+		Page<ProductSearchDto> result;
+		try {
+			result = productService.findProductsByDynamicConditions(categoryEnum,
+				productType, minPrice, maxPrice, locations, titleKeyword, pageable);
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.INTERNAL_SERVER_ERROR);
+		}
 
 		return ApiResponse.createSuccess(result, "성공적으로 조회하였습니다.");
 	}
